+++ src/translations/components/TranslationsShippingMethodPage/TranslationsShippingMethodPage.tsx
@@ -4,7 +4,10 @@ import LanguageSwitch from "@saleor/components/LanguageSwitch";
 import PageHeader from "@saleor/components/PageHeader";
 import { ShippingMethodTranslationFragment } from "@saleor/fragments/types/ShippingMethodTranslationFragment";
 import { commonMessages, sectionNames } from "@saleor/intl";
-import { TranslationsEntitiesPageProps } from "@saleor/translations/types";
+import {
+  TranslationInputFieldName,
+  TranslationsEntitiesPageProps
+} from "@saleor/translations/types";
 import React from "react";
 import { useIntl } from "react-intl";
 
@@ -15,10 +18,6 @@ export interface TranslationsShippingMethodPageProps
   extends TranslationsEntitiesPageProps {
   data: ShippingMethodTranslationFragment;
 }
-
-export const fieldNames = {
-  name: "name"
-};
 
 const TranslationsShippingMethodPage: React.FC<TranslationsShippingMethodPageProps> = ({
   activeField,
@@ -67,12 +66,23 @@ const TranslationsShippingMethodPage: React.FC<TranslationsShippingMethodPagePro
         fields={[
           {
             displayName: intl.formatMessage({
-              defaultMessage: "ShippingMethod Name"
+              defaultMessage: "Name",
+              description: "shipping method name"
             }),
-            name: fieldNames.name,
+            name: TranslationInputFieldName.name,
             translation: data?.translation?.name || null,
             type: "short" as "short",
             value: data?.name
+          },
+          {
+            displayName: intl.formatMessage({
+              defaultMessage: "Description",
+              description: "shipping method description"
+            }),
+            name: TranslationInputFieldName.description,
+            translation: data?.translation?.description || null,
+            type: "rich",
+            value: data?.description
           }
         ]}
         saveButtonState={saveButtonState}